PP-3603 - refactoring old style controllers
