+++ app/assets/sass/modules/_service_switcher.scss
@@ -42,7 +42,7 @@
       position: absolute;
       top: 50%;
       left: 0;
-      @include translate(0, -60%);
+      @include translate(0, -50%);
       width: 1.5rem;
       height: 1.5rem;
       border-radius: 100%;
