+++ app/controllers/my-services/index.js
@@ -1,0 +1,4 @@
+'use strict'
+
+exports.getIndex = require('./get-index-controller')
+exports.postIndex = require('./post-index-controller')
